+++ R/bp_distribution.R
@@ -2,6 +2,38 @@
 #'
 #' Percentile and quantile functions for pediatric blood pressure.
 #'
+#' \code{source} is used to specify the method or source data sets by which the
+#' percentiles are generated.  This can be controlled by the option
+#' \code{pedbp_bp_source}.
+#' End users are encouraged to set the option if not using the default so all
+#' calls to these functions will use the same source.
+#'
+#' Options:
+#' \itemize{
+#'   \item \code{martin2022} (default) uses a combination of references to generate
+#'   percentiles ages from 1 months through 18 years, without or without known
+#'   stature.  This was the only method implemented in version 1 of the pedbp package.
+#'
+#'   \item \code{gemelli1990} uses only the reference values from Gemelli et al.
+#'   (1990).  These values are applicable to patients from 1 month to 12 months
+#'   of age. Stature is not used in the look up for the parameters.
+#'
+#'   \item \code{lo2013} uses only the reference values from Lo et al. (2013).
+#'   This is applicable to patients of at least three years of age.  Height is
+#'   not considered when looking up the parameters.
+#'
+#'   \item \code{nhlbi} uses only reference values from the National Heart,
+#'   Lung, and Blood Institute [NHLBI] and the Centers for Disease Control and
+#'   Prevention [CDC] published in 2011.  These are for patients of at least one
+#'   year of age and with a known stature.  These values were publish
+#'
+#'   \item \code{flynn2017} uses only reference values from Flynn et al. (2017).
+#'   These values are similar to the nhlbi values _but_ "do not include children
+#'   and adolescents with overweight and obesity (ie, those with a BMI >= 85th
+#'   percentile).
+#' }
+#'
+#'
 #' @param q_sbp a vector of systolic blood pressures
 #' @param q_dbp a vector of diastolic blood pressures
 #' @param p_sbp a vector of systolic blood percentiles
@@ -11,6 +43,7 @@
 #' @param height numeric, in centimeters, can be missing.
 #' @param height_percentile default height percentile to use if \code{height} is
 #' missing.
+#' @param source the method, or data set, to use as the reference.  See Details.
 #' @param ... not currently used
 #'
 #' @seealso \code{vignette("bp-distriution", package = "pedbp")}
@@ -20,6 +53,31 @@
 #' (diastolic blood pressure).  Additionally, the \code{bp_params} attribute
 #' provides details on the data source and parameters used in the percentile
 #' estimates.
+#'
+#' @references
+#'
+#' Gemelli, Marina, Rosa Manganaro, Carmelo Mamì, and F. De Luca. "Longitudinal
+#' study of blood pressure during the 1st year of life." European journal of
+#' pediatrics 149 (1990): 318-320.
+#'
+#' Lo, Joan C., Alan Sinaiko, Malini Chandra, Matthew F. Daley, Louise C.
+#' Greenspan, Emily D. Parker, Elyse O. Kharbanda et al. "Prehypertension and
+#' hypertension in community-based pediatric practice." Pediatrics 131, no. 2
+#' (2013): e415-e424.
+#'
+#' "Expert panel on integrated guidelines for cardiovascular health and risk
+#' reduction in children and adolescents: summary report." Pediatrics 128, no.
+#' Suppl 5 (2011): S213. <doi:10.1542/peds.2009-2107C>
+#'
+#' The Fourth Report on the Diagnosis, Evaluation, and Treatment of High Blood
+#' Pressure in Children and Adolescents National High Blood Pressure Education
+#' Program Working Group on High Blood Pressure in Children and Adolescents
+#' Pediatrics 2004;114;555-576 <doi:10.1542/peds.114.2.S2.555>
+#'
+#' Flynn, Joseph T., David C. Kaelber, Carissa M. Baker-Smith, Douglas Blowey,
+#' Aaron E. Carroll, Stephen R. Daniels, Sarah D. De Ferranti et al. "Clinical
+#' practice guideline for screening and management of high blood pressure in
+#' children and adolescents." Pediatrics 140, no. 3 (2017).
 #'
 #' @examples
 #'
@@ -68,13 +126,28 @@
 #'   , male  = d$male
 #'   )
 #'
+#' #############################################################################
+#' # Selecting different source values
+#'
+#' p_bp(q_sbp = 92, q_dbp = 60, age = 29.2, male = 0, height_percentile = 0.95, source = "martin2022") # defualt
+#' p_bp(q_sbp = 92, q_dbp = 60, age = 29.2, male = 0, height_percentile = 0.95, source = "gemelli1990")
+#' p_bp(q_sbp = 92, q_dbp = 60, age = 29.2, male = 0, height_percentile = 0.95, source = "lo2013")
+#' p_bp(q_sbp = 92, q_dbp = 60, age = 29.2, male = 0, height_percentile = 0.95, source = "nhlbi")
+#' p_bp(q_sbp = 92, q_dbp = 60, age = 29.2, male = 0, height_percentile = 0.95, source = "flynn2017")
+#'
+#' q_bp(p_sbp = 0.85, p_dbp = 0.85, age = 29.2, male = 0, height_percentile = 0.95, source = "martin2022") # defualt
+#' q_bp(p_sbp = 0.85, p_dbp = 0.85, age = 29.2, male = 0, height_percentile = 0.95, source = "gemelli1990")
+#' q_bp(p_sbp = 0.85, p_dbp = 0.85, age = 29.2, male = 0, height_percentile = 0.95, source = "lo2013")
+#' q_bp(p_sbp = 0.85, p_dbp = 0.85, age = 29.2, male = 0, height_percentile = 0.95, source = "nhlbi")
+#' q_bp(p_sbp = 0.85, p_dbp = 0.85, age = 29.2, male = 0, height_percentile = 0.95, source = "flynn2017")
+#'
 #'
 #' @name bp_distribution
 NULL
 
 #' @rdname bp_distribution
 #' @export
-p_bp <- function(q_sbp, q_dbp, age, male, height = NA, height_percentile = 0.50, ...) {
+p_bp <- function(q_sbp, q_dbp, age, male, height = NA, height_percentile = 0.50, source = getOption("pedbp_bp_source", "martin2022"), ...) {
 
   stopifnot(length(q_sbp) == length(q_dbp))
 
@@ -102,7 +175,7 @@ p_bp <- function(q_sbp, q_dbp, age, male, height = NA, height_percentile = 0.50,
     stop("length(height) needs to be at least 1L")
   }
 
-  d <- v_bp_params(age = age, male = male, height = height, height_percentile = height_percentile)
+  d <- v_bp_params(age = age, male = male, height = height, height_percentile = height_percentile, source = source)
 
   rtn <-
     list(sbp_percentile = stats::pnorm(q_sbp, mean = d$sbp_mean, sd = d$sbp_sd)
@@ -114,7 +187,7 @@ p_bp <- function(q_sbp, q_dbp, age, male, height = NA, height_percentile = 0.50,
 
 #' @rdname bp_distribution
 #' @export
-q_bp <- function(p_sbp, p_dbp, age, male, height = NA, height_percentile = 0.50, ...) {
+q_bp <- function(p_sbp, p_dbp, age, male, height = NA, height_percentile = 0.50, source = getOption("pedbp_bp_source", "martin2022"), ...) {
 
   stopifnot(length(p_sbp) == length(p_dbp))
 
@@ -142,7 +215,7 @@ q_bp <- function(p_sbp, p_dbp, age, male, height = NA, height_percentile = 0.50,
     stop("length(height) needs to be at least 1L")
   }
 
-  d <- v_bp_params(age = age, male = male, height = height, height_percentile = height_percentile)
+  d <- v_bp_params(age = age, male = male, height = height, height_percentile = height_percentile, source = source)
 
   rtn <-
     list(sbp = stats::qnorm(p_sbp, mean = d$sbp_mean, sd = d$sbp_sd)
@@ -163,22 +236,38 @@ print.pedbp_bp <- function(x, ...) {
 # was easier to build assuming age, male, height were single values. v_bp_params
 # will make functions easier to use for end users.
 
-v_bp_params <- function(age, male, height, height_percentile = 0.50, ...) {
+v_bp_params <- function(age, male, height, height_percentile = 0.50, source = getOption("pedbp_bp_source", "martin2022"), ...) {
 
-  rtn <- Map(bp_params, age = age, male = male, height = height, height_percentile = height_percentile)
+  rtn <- Map(bp_params, age = age, male = male, height = height, height_percentile = height_percentile, source = source)
   do.call(rbind, rtn)
 
 }
 
 
-bp_params <- function(age, male, height = NA, height_percentile = 0.50, ...) {
+bp_params <- function(age, male, height = NA, height_percentile = 0.50, source = getOption("pedbp_bp_source", "martin2022"), ...) {
+
   stopifnot(length(age) == 1L)
   stopifnot(length(male) == 1L)
-  stopifnot(all(age >=0) & all(age < 19 * 12))
-  stopifnot(all(male %in% c(0L, 1L)))
-  stopifnot(all(stats::na.omit(height > 0)))
+  #stopifnot(all(age >=0))# & all(age < 19 * 12))
+  #stopifnot(all(male %in% c(0L, 1L)))
+  #stopifnot(all(stats::na.omit(height > 0)))
   stopifnot(0 < height_percentile & height_percentile < 1)
 
+  source <- match.arg(source, choices = c("martin2022", "gemelli1990", "nhlbi", "lo2013", "flynn2017"), several.ok = FALSE)
+
+  e <- new.env()
+  utils::data(list = "bp_parameters", package = "pedbp", envir = e)
+  d <- e$bp_parameters[e$bp_parameters$male == male, ]
+
+  if (age < 0 | age >= 19*12) {
+    d <- e$bp_parameters[1, ]
+    for (j in names(d)) {
+      d[j] <- NA
+    }
+  }
+
+  # get the height_percentile if height is not NA
+  # used for source %in% c("martin2022", "flynn2017")
   if (!is.na(height)) {
     if (age < 36) {
       height_percentile <- p_length_for_age(q = height, age = age, male = male, source = "WHO")
@@ -187,27 +276,58 @@ bp_params <- function(age, male, height = NA, height_percentile = 0.50, ...) {
     }
   }
 
-  e <- new.env()
-  utils::data(list = "bp_parameters", package = "pedbp", envir = e)
-  d <- e$bp_parameters[e$bp_parameters$male == male, ]
 
-  if (age < 12) {
-    d <- d[d$age == min(d$age) | d$age <= age, ]
-    d <- d[d$age == max(d$age), ]
-    height_percentile <- NA_real_
-  } else if (is.na(height) & age >= 36) {
-    d <- d[is.na(d$height_percentile), ]
-    d <- d[((age >= 36) & (d$age <= age)), ]
-    d <- d[d$age == max(d$age), ]
-    height_percentile <- NA_real_
-  } else {
-    d <- d[!is.na(d$height_percentile), ]
+  if (source == "martin2022") {
+    d <- d[d$source %in% c("gemelli1990", "lo2013", "nhlbi"), ]
+    if (age < 12) {
+      d <- d[d$age <= age, ]
+      if (nrow(d) > 1L) {
+        d <- d[d$age == max(d$age), ]
+      }
+    } else if (is.na(height) & age >= 36) {
+      d <- d[is.na(d$height_percentile), ]
+      d <- d[((age >= 36) & (d$age <= age)), ]
+      if (nrow(d) > 1L) {
+        d <- d[d$age == max(d$age), ]
+      }
+    } else {
+      d <- d[!is.na(d$height_percentile), ]
+      d <- d[d$age <= age, ]
+      if (nrow(d) > 1L) {
+        d <- d[d$age == max(d$age), ]
+      }
+      d <- d[which.min(abs(d$height_percentile/100 - height_percentile)), ]
+    }
+  } else if (source %in% c("gemelli1990", "lo2013")) {
+    if (source == "gemelli1990" & age > 12) {
+      d <- d[0, ]
+    } else {
+      d <- d[d$source == source, ]
+      d <- d[d$age <= age, ]
+      if (nrow(d) > 1L) {
+        d <- d[d$age == max(d$age), ]
+      }
+    }
+  } else if (source %in% c("nhlbi", "flynn2017")) {
+    d <- d[d$source == source, ]
     d <- d[d$age <= age, ]
-    d <- d[d$age == max(d$age), ]
+    if (nrow(d) > 1L) {
+      d <- d[d$age == max(d$age), ]
+    }
     d <- d[which.min(abs(d$height_percentile/100 - height_percentile)), ]
+  } else {
+    stop("unknown source")
+  }
+
+  if (nrow(d) == 0L) {
+    d <- e$bp_parameters[1, ]
+    for (j in names(d)) {
+      d[[j]] <- NA
+    }
   }
 
   stopifnot(nrow(d) == 1L)
+
   d
 }