+++ tests/test-bp_parameters.R
@@ -6,12 +6,17 @@ library(pedbp)
 stopifnot(identical(class(bp_parameters), "data.frame"))
 
 # verify some basic features of the data.frame
-stopifnot(identical(nrow(bp_parameters), 276L))
 stopifnot(identical(ncol(bp_parameters),   8L))
+
+# rows for source == 'martin2022'
+stopifnot(identical(nrow(bp_parameters[bp_parameters$source %in% c('gemelli1990', 'lo2013', 'nhlbi'), ]), 276L))
+
+# total rows
+stopifnot(identical(nrow(bp_parameters), 514L))
 
 stopifnot(identical(
           sort(unique(bp_parameters[["source"]]))
-          , c("gemelli1990", "lo2013", "nhlbi")
+          , c("flynn2017", "gemelli1990", "lo2013", "nhlbi")
           ))
 
 stopifnot(identical(